feat: well tests and some refactoring to match tests

+++ PRIO/DTOS/WellDTOS/WellHistoryDTO.cs
@@ -1,8 +1,5 @@
 ﻿using PRIO.DTOS.FieldDTOS;
 using PRIO.DTOS.UserDTOS;
-using PRIO.Models.Fields;
-using PRIO.Models.Users;
-using PRIO.Models.Wells;
 
 namespace PRIO.DTOS.WellDTOS
 {
@@ -26,12 +23,12 @@ namespace PRIO.DTOS.WellDTOS
         public bool? StatusOperatorOld { get; set; }
         public string? Type { get; set; }
         public string? TypeOld { get; set; }
-        public double? WaterDepth { get; set; }
-        public double? WaterDepthOld { get; set; }
-        public double? TopOfPerforated { get; set; }
-        public double? TopOfPerforatedOld { get; set; }
-        public double? BaseOfPerforated { get; set; }
-        public double? BaseOfPerforatedOld { get; set; }
+        public decimal? WaterDepth { get; set; }
+        public decimal? WaterDepthOld { get; set; }
+        public decimal? TopOfPerforated { get; set; }
+        public decimal? TopOfPerforatedOld { get; set; }
+        public decimal? BaseOfPerforated { get; set; }
+        public decimal? BaseOfPerforatedOld { get; set; }
         public string? ArtificialLift { get; set; }
         public string? ArtificialLiftOld { get; set; }
         public string? Latitude4C { get; set; }
